ccsovs/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-background-tasks==1.2.5
django-bootstrap-v5==1.0.7
django-common-helpers==0.9.2
django-compat==1.0.15
django-crispy-forms==1.13.0
django-cron==0.5.1
django-filter==21.1
django-jquery==3.1.0
django-tinymce==3.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
Markdown==3.3.4
openpyxl==3.0.9
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
